feat: add tracking to offline search completions
