+++ LegalLead.PublicData.Search/Helpers/GenExcelFileParameter.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Thompson.RecordSearch.Utility.Models;
 
 namespace LegalLead.PublicData.Search.Helpers
 {
@@ -11,5 +12,18 @@ namespace LegalLead.PublicData.Search.Helpers
         public string TrackingIndex { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
+        public int RecordCount { get; set; }
+        public string FileName { get; set; }
+
+
+        public string GetShortName()
+        {
+            if (string.IsNullOrWhiteSpace(FileName)) return string.Empty;
+            var fileName = System.IO.Path.GetFileNameWithoutExtension(FileName);
+            var extn = System.IO.Path.GetExtension(FileName);
+            if (string.IsNullOrWhiteSpace(fileName)) return FileName;
+            if (string.IsNullOrWhiteSpace(extn)) extn = string.Empty;
+            return string.Concat(fileName, extn);
+        }
     }
 }
