+++ LegalLead.PublicData.Search/LegalLead.PublicData.Search.csproj
@@ -210,6 +210,9 @@
     <Content Include="FsInvoiceHistory.ProcessBackground.cs">
       <DependentUpon>FsInvoiceHistory.cs</DependentUpon>
     </Content>
+    <Content Include="FsOfflineHistory.Models.cs">
+      <DependentUpon>FsOfflineHistory.cs</DependentUpon>
+    </Content>
   </ItemGroup>
   <ItemGroup Label="Framework48">
     <Compile Remove="obj\Integration\.NETFramework,Version=v4.8.AssemblyAttributes.cs" />
